flatten identity mod
