--- xmtp_mls/src/identity/mod.rs
@@ -1,1 +1,0 @@
-pub mod xmtp_id;
