+++ xmtp_mls/src/storage/mod.rs
@@ -4,8 +4,7 @@ mod serialization;
 pub mod sql_key_store;
 
 pub use encrypted_store::{
-    db_connection, group, group_intent, group_message, identity, identity_update,
-    key_store_entry, refresh_state, EncryptedMessageStore, EncryptionKey, RawDbConnection,
-    StorageOption,
+    db_connection, group, group_intent, group_message, identity, identity_update, key_store_entry,
+    refresh_state, EncryptedMessageStore, EncryptionKey, RawDbConnection, StorageOption,
 };
 pub use errors::StorageError;
